lib/simbricks: big refactor of protocol and API
High-level idea is to separate a out a base-protocol and library that
the PCIe and network protocols layer over. The base-protocol includes
synchronization and the basics of setting up queues, and transferring
messages. So this will enable us to build protocol-agnostic pieces like
proxies, loggers etc. A secondary goal down the line is to remove the
fixed roles of  "nics listen on pcie & network" while hosts and networks
connect, instead making this dynamic.

Unfortunately this is a big change touching just about every piece
interfacing with simbricks. So this is also consolidating and
restructuring our simbricks libraries and adapters, reducing copying of
code etc. After this commit all simulators, ours and external ones, are
broken till the respective commits fixing them.