orchestration: use echo as default checksum cmd
Currently we invoke m5 checksum for gem5 but nothing at all for other
simulators. This just adds an echo "ready for checkpoint". We use this in
simics to generate a checkpoint, but it is also sane default for simulators not
supporting checkpoints, as at least the logs will show what's going on.