orchestration/simulators.py: speed up simulations using gem5 checkpoints
Currently, when a checkpoint is created in gem5 using the KVM CPU, the
simulated tick when the checkpoint is taken will end up in the few
seconds or more when the KVM CPU is slow. When the checkpoint is
afterwards restored, gem5 will continue the simulation at this tick.
This means that other, synchronized simulators first have to catch up,
essentially just synchronizing without performing any useful work.
Depending on their speed, this can take a lot of time and is especially
costly with a RTL simulator in the loop. To avoid this we now, by
default, edit the to be restored gem5 checkpoint to start at tick 0
instead.