devcontainer: fix `make documentation` error due to missing locale
VS Code started enforcing the locale but they decided to use en_US.UTF-8. It looks like they are going to change this to C.UTF-8, which is available on every POSIX compliant system but this might take some time. So instead, we add the required en_US.UTF-8 locale to the devcontainer for now.